+++ shipping_microservice/main/services/__init__.py
@@ -1,2 +1,2 @@
-from .stock_service import StockService
+from .order_service import OrderService
 from .ship_service import ShipService
