+++ shipping_microservice/main/services/order_service.py
@@ -1,10 +1,12 @@
-from main.schemas import ArticlesOrderSchema
-from main.repositories import ArticlesOrderRepository
+from main.schemas import ArticlesOrderSchema, OrderSchema
+from main.repositories import ArticlesOrderRepository, OrderRepository
 
 articlesorder_schema = ArticlesOrderSchema()
 articlesorder_repository = ArticlesOrderRepository()
+order_repository = OrderRepository()
+order_schema = OrderSchema()
 
-class StockService:
+class OrderService:
 
     #articlesorder
     def add_articlesorder(self, articlesorder):
@@ -21,4 +23,11 @@ class StockService:
     def delete_articlesorder(self, articlesorder):
         articlesorder_repository.delete(articlesorder)
         return articlesorder_schema.dump(articlesorder)
+    
+    def add_order(self, order):
+        order_repository.create(order)
+        return order_schema.dump(order)
+    
+    def get_order(self, id):
+        return order_repository.find_by_id(id)
